requirements.txt
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
argon2-cffi==20.1.0
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.4
backcall==0.2.0
BaseHash==3.0.8
beautifulsoup4==4.9.3
billiard==3.3.0.23
blessings==1.7
boto3==1.16.40
botocore==1.19.40
bpython==0.20.1
CacheControl==0.12.6
celery==3.1.26.post2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
colorama==0.4.4
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
curtsies==0.3.4
decorator==4.4.2
distlib==0.3.0
distro==1.4.0
Django==3.1.4
django-anymail==8.1
django-bootstrap4==2.3.1
django-celery==3.3.1
django-ckeditor==6.0.0
django-cors-headers==3.6.0
django-extensions==3.1.0
django-filter==2.4.0
django-js-asset==1.2.2
django-rest-swagger==2.2.0
django-silk==4.1.0
django-storages==1.11
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
docutils==0.16
factory-boy==3.2.0
Faker==5.4.0
gprof2dot==2019.11.30
greenlet==0.4.17
html5lib==1.0.1
idna==2.10
ipaddr==2.2.0
ipython==7.19.0
ipython-genutils==0.2.0
isort==5.7.0
itypes==1.2.0
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
kombu==3.0.37
lazy-object-proxy==1.4.3
lockfile==0.12.2
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
msgpack==0.6.2
openapi-codec==1.3.2
packaging==20.3
parso==0.7.1
pep517==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
progress==1.5
prompt-toolkit==3.0.8
ptyprocess==0.7.0
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.7.3
PyJWT==1.7.1
pylint==2.6.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-decouple==3.3
pytoml==0.1.21
pytz==2020.4
requests==2.25.1
retrying==1.3.3
s3transfer==0.3.3
sentry-sdk==0.19.5
simplejson==3.17.2
six==1.15.0
snakeviz==2.1.0
soupsieve==2.1
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tornado==6.1
traitlets==5.0.5
uritemplate==3.0.1
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
